+++ packages/larva-patterns/modules/article-kicker/article-kicker.prototype.js
@@ -1,0 +1,6 @@
+module.exports = {
+	article_kicker_classes: '',
+	article_kicker_text: 'Article Kicker',
+	article_kicker_url: '#',
+	article_kicker_link_classes: '',
+};