+++ packages/larva-patterns/modules/article-title/article-title.prototype.js
@@ -1,0 +1,4 @@
+module.exports = {
+	article_title_classes: '',
+	article_title_markup: 'Frieze Editorial Director to Step Down, and More from September 6, 2019',
+};